// src/utils/foodCategories/types.ts
// Predefined food categories in German
export enum FoodCategory {
  FRUECHTE = 'Früchte',
  GEMUESE = 'Gemüse',
  FLEISCH = 'Fleisch',
  FISCH = 'Fisch',
  MILCHPRODUKTE = 'Milchprodukte',
  GETREIDE = 'Getreide & Backwaren',
  HUELSENFRUECHTE = 'Hülsenfrüchte',
  NUESSE = 'Nüsse & Samen',
  GEWUERZE = 'Gewürze',
  GETRÄNKE = 'Getränke',
  SUESSIGKEITEN = 'Süßigkeiten',
  KONSERVEN = 'Konserven',
  TIEFKUEHLWARE = 'Tiefkühlware',
  SONSTIGES = 'Sonstiges'
}